Changed 'result' field in as_transaction from int to uint8_t.

+++ as/include/base/secondary_index.h
@@ -540,7 +540,7 @@ void as_sindex_putall_rd(as_namespace *ns, as_storage_rd *rd);
 // **************************************************************************************************
 extern int                  as_sindex_ns_has_sindex(as_namespace *ns);
 extern const char         * as_sindex_err_str(int err_code);
-extern int                  as_sindex_err_to_clienterr(int err, char *fname, int lineno);
+extern uint8_t              as_sindex_err_to_clienterr(int err, char *fname, int lineno);
 extern bool                 as_sindex_isactive(as_sindex *si);
 extern int                  as_sindex_get_err(int op_code, char *filename, int lineno);
 extern uint64_t             as_sindex_get_ns_memory_used(as_namespace *ns);
